feat/receipt : move folder
